+++ src/electron/adapters/direct-action-message-dispatcher.ts
